Improve the ALIGN_STACK_PTR macro
Now casts to size_t for you.

Also add alignments for Thumb and Arm. Which aren't
needed right now but will be for some later work.
(Thumb could be 4 or 8 but chose 8 for simplicity)

Also removed comments calling out AArch64 specifically
since we're aligning on all platforms now.

+++ include/port/aarch64.h
@@ -78,8 +78,9 @@ typedef struct {
   size_t pc;
 } __attribute__((packed)) FiberContext;
 
-#define PC_ADD_MODE(pc)      (pc)
-#define PC_REMOVE_MODE(pc)   (pc)
-#define ALIGN_STACK_PTR(ptr) ((ptr) & ~0xF)
+#define PC_ADD_MODE(pc)    (pc)
+#define PC_REMOVE_MODE(pc) (pc)
+// 16 byte alignment
+#define ALIGN_STACK_PTR(ptr) ((size_t)(ptr) & ~0xF)
 
 #endif /* ifdef PORT_AARCH64_H */
